web-form-ui-field-detection/requirements.txt
flask>=3.0
ultralytics>=8.3.0
pillow>=10.0
huggingface_hub>=0.20
